Remove unnecessary `source` from the translation function
I had an idea that specifying "source" would help GPT to understand that it would need to find the source to translate in the prompt and also partly for debugging purposes, to see that it read/parsed the source correctly. I don't know if this was a good idea or not. Now it appeard to just cause it confusion, it sometimes provided responses that was missing the `translation` property and it had instead put the translated text in the source property. So might be better to remove it.